Updated delete function to only be possible if token matches blog post creator

+++ BlogList/controllers/blogs.js
@@ -34,8 +34,20 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const targetBlog = await Blog.findById(request.params.id)
+
+  if (targetBlog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'incorrect user'})
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
